Make LineFormatter abstract
