+++ src/main/scala/wdlTools/formatter/DefaultLineFormatter.scala
@@ -1,0 +1,138 @@
+package wdlTools.formatter
+
+import wdlTools.formatter.Indenting.Indenting
+import wdlTools.formatter.Wrapping.Wrapping
+
+import scala.collection.mutable
+
+case class DefaultLineFormatter(defaultIndenting: Indenting = Indenting.IfNotIndented,
+                                initialIndent: String = "",
+                                indentation: String = " ",
+                                indentStep: Int = 2,
+                                spacing: String = " ",
+                                maxLineWidth: Int = 100,
+                                lines: mutable.Buffer[String] = mutable.ArrayBuffer.empty)
+    extends LineFormatter(defaultIndenting) {
+  private val currentLine: mutable.StringBuilder = new StringBuilder(maxLineWidth)
+  private val indent: mutable.StringBuilder = new mutable.StringBuilder(initialIndent)
+
+  def indented(indenting: Indenting = defaultIndenting): LineFormatter = {
+    DefaultLineFormatter(
+        defaultIndenting = indenting,
+        initialIndent = initialIndent + (indentation * indentStep),
+        indentation = indentation,
+        indentStep = indentStep,
+        maxLineWidth = maxLineWidth,
+        lines = lines
+    )
+  }
+
+  def preformatted(): LineFormatter = {
+    DefaultLineFormatter(
+        spacing = "",
+        maxLineWidth = maxLineWidth,
+        lines = lines
+    )
+  }
+
+  def atLineStart: Boolean = {
+    currentLine.length <= indent.length
+  }
+
+  def lengthRemaining: Int = {
+    if (atLineStart) {
+      maxLineWidth
+    } else {
+      maxLineWidth - currentLine.length
+    }
+  }
+
+  def dent(indenting: Indenting = defaultIndenting): Unit = {
+    indenting match {
+      case Indenting.Always => indent.append(indentation * indentStep)
+      case Indenting.IfNotIndented if indent.length == initialIndent.length =>
+        indent.append(indentation * indentStep)
+      case Indenting.Dedent =>
+        val indentLength = indent.length
+        if (indentLength > initialIndent.length) {
+          indent.delete(indentLength - indentStep, indentLength)
+        }
+      case Indenting.Reset =>
+        indent.clear()
+        indent.append(initialIndent)
+      case _ => Unit
+    }
+  }
+
+  def emptyLine(): Unit = {
+    lines.append("")
+  }
+
+  def beginLine(): Unit = {
+    require(atLineStart)
+    if (currentLine.isEmpty) {
+      currentLine.append(indent)
+    }
+  }
+
+  def endLine(wrap: Boolean = false, indenting: Indenting = defaultIndenting): Unit = {
+    if (!atLineStart) {
+      lines.append(currentLine.toString)
+      if (wrap) {
+        dent(indenting)
+      } else {
+        dent(Indenting.Reset)
+      }
+    }
+    currentLine.clear()
+  }
+
+  def buildSubstring(
+      chunks: Seq[Chunk],
+      builder: mutable.StringBuilder = new mutable.StringBuilder(maxLineWidth)
+  ): StringBuilder = {
+    chunks.foreach { chunk =>
+      if (builder.nonEmpty && !(builder.last.isWhitespace || builder.last == indentation.last)) {
+        builder.append(spacing)
+      }
+      builder.append(chunk.toString)
+    }
+    builder
+  }
+
+  def appendString(value: String): Unit = {
+    currentLine.append(value)
+  }
+
+  def appendChunk(chunk: Chunk): Unit = {
+    buildSubstring(Vector(chunk), currentLine)
+  }
+
+  def appendAll(chunks: Seq[Chunk], wrapping: Wrapping = Wrapping.AsNeeded): Unit = {
+    if (wrapping == Wrapping.Never) {
+      buildSubstring(chunks, currentLine)
+    } else {
+      val substr = buildSubstring(chunks)
+      if (wrapping == Wrapping.Always) {
+        endLine(wrap = true)
+      }
+      val space = if (atLineStart) {
+        ""
+      } else {
+        spacing
+      }
+      if (wrapping != Wrapping.Never && lengthRemaining < space.length + substr.length) {
+        chunks.foreach { chunk =>
+          chunk.format(lineFormatter = this)
+        }
+      } else {
+        currentLine.append(space)
+        currentLine.append(substr)
+      }
+    }
+  }
+
+  def toSeq: Seq[String] = {
+    lines.toVector
+  }
+}
