+++ src/main/scala/wdlTools/formatter/WDL10Formatter.scala
@@ -745,10 +745,10 @@ case class WDL10Formatter(verbosity: Verbosity = Verbosity.Normal,
         println(Util.prettyFormat(statements))
       }
 
-      val lineFormatter = LineFormatter()
+      val lineFormatter = DefaultLineFormatter()
       format(lineFormatter)
 
-      documents(uri) = lineFormatter.lines.toVector
+      documents(uri) = lineFormatter.toSeq
     }
   }
